requirements.txt
argparse
flask[async]
fileloghelper
httpx>=0.18,<0.24
pytest
pytest-cov
pytest-asyncio
pytest-httpx
pytest-flask
pyfakefs
pylint
pylint-exit
aiohttp
mypy
mypy-extensions
types-requests
tox==3.*
python-crontab
croniter
watchdog
pid
docker
gunicorn
requests
semver
GitPython
moodle-dl
pyyaml
kubernetes
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
